pyproject.toml
[project]
name = "liam-rs"
version = "0.0.0"
description = "Scripts for my personal website"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["python-frontmatter>=1.1.0"]

[dependency-groups]
dev = ["ruff>=0.7.0"]

[tool.ruff]
src = ["src"]
indent-width = 2
line-length = 100

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 20
indent-style = "space"
quote-style = "single"
